requirements.txt
streamlit==1.32.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
sentence-transformers>=2.2.2
Pillow>=9.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
requests>=2.28.0
accelerate>=0.21.0
diffusers>=0.25.0
protobuf>=3.20.0
sentencepiece>=0.1.99
groq
